+++ Server/Cinema/CinemaApp.Application/DTOs/Seat/UpdateSeatDto.cs
@@ -11,7 +11,7 @@ namespace CinemaApp.Application.DTOs.Seat
         public string RowName { get; set; }
 
         [Required]
-        public short SeatNo { get; set; }
+        public int SeatNo { get; set; }
 
         [Required]
         public int SeatGroupId { get; set; }
