+++ Server/Cinema/CinemaApp.Domain/Entities/Seat.cs
@@ -9,7 +9,7 @@
         public int SeatGroupId { get; set; }
         public int IndexInsideSeatGroup { get; set; }
         public string RowName { get; set; }
-        public short SeatNo { get; set; }
+        public int SeatNo { get; set; }
         public SeatType SeatType { get; set; }
         public ICollection<Ticket> Tickets { get; set; }
         public ICollection<TicketSeat> TicketSeats { get; set; }
