+++ Server/Cinema/CinemaApp.Infrastructure/Data/Tables/CreateSeatsTable.sql
@@ -6,7 +6,7 @@
     IndexInsideSeatGroup INT                NOT NULL,
     SeatTypeId           INT                    NULL,
     RowName              VARCHAR(15)        NOT NULL,
-    SeatNo               SMALLINT           NOT NULL,
+    SeatNo               INT                NOT NULL,
     CONSTRAINT PK_Seats PRIMARY KEY CLUSTERED (Id ASC),
     CONSTRAINT FK_Seats_Halls FOREIGN KEY (HallId) REFERENCES Halls (Id) ON DELETE CASCADE,
     CONSTRAINT FK_Seats_SeatTypes FOREIGN KEY (SeatTypeId) REFERENCES SeatTypes (Id) ON DELETE SET NULL
